major improvements to file org - authwrapper
