+++ src/authwrapp/sign-in/index.js
@@ -1,0 +1,1 @@
+export { default as default } from './container';
